+++ src/hapi/pipelines/database/operational_presence.py
@@ -3,6 +3,7 @@ from logging import getLogger
 from typing import Dict
 
 from hapi_schema.db_operational_presence import DBOperationalPresence
+from hdx.location.names import clean_name
 from sqlalchemy.orm import Session
 
 from . import admins
@@ -56,26 +57,42 @@ class OperationalPresence(BaseUploader):
 
                 for admin_code, org_names in values[org_name_index].items():
                     for i, org_name in enumerate(org_names):
-                        org_acronym = values[org_acronym_index][admin_code][i]
-                        org_type_code = None
-                        if org_type_name_index:
-                            org_type_name = values[org_type_name_index][
-                                admin_code
-                            ][i]
-                            org_type_code = self._org_type.get_org_type_code(
-                                org_type_name
-                            )
-                            if not org_type_code:
-                                logger.error(
-                                    f"Org type {org_type_name} not in table"
+                        # TODO: find the country code for get_org_info parameter "location"
+                        org_info = self._org.get_org_info(
+                            org_name, location="Country code"
+                        )
+                        self._org.add_org_to_lookup(
+                            org_name, org_info.get("#org+name")
+                        )
+                        org_name = org_info.get("#org+name")
+                        org_acronym = org_info.get(
+                            "#org+acronym",
+                            values[org_acronym_index][admin_code][i],
+                        )
+                        org_type_code = org_info.get("#org+type+code")
+                        if not org_type_code:
+                            if org_type_name_index:
+                                org_type_name = values[org_type_name_index][
+                                    admin_code
+                                ][i]
+                                org_type_code = (
+                                    self._org_type.get_org_type_code(
+                                        org_type_name
+                                    )
                                 )
+                        if not org_type_code:
+                            logger.error("Org type missing or not in table")
                         # TODO: find out how unique orgs are. Currently checking that
                         #  combo of acronym/name/type is unique. (More clarity will come
                         #  from HAPI-166).
                         if (
                             org_acronym is not None
                             and org_name is not None
-                            and (org_acronym, org_name, org_type_code)
+                            and (
+                                org_acronym.upper(),
+                                clean_name(org_name),
+                                org_type_code,
+                            )
                             not in self._org.data
                         ):
                             self._org.populate_single(
@@ -113,7 +130,11 @@ class OperationalPresence(BaseUploader):
                             resource_id
                         ]
                         org_ref = self._org.data[
-                            (org_acronym, org_name, org_type_code)
+                            (
+                                org_acronym.upper(),
+                                clean_name(org_name),
+                                org_type_code,
+                            )
                         ]
                         admin2_ref = self._admins.admin2_data[admin2_code]
                         row = (resource_ref, org_ref, sector_code, admin2_ref)